+++ packages/editor/src/components/post-publish-panel/prepublish.js
@@ -14,7 +14,6 @@ import MaybePostFormatPanel from './maybe-post-format-panel';
 import { store as editorStore } from '../../store';
 import MaybeCategoryPanel from './maybe-category-panel';
 import MaybeUploadMedia from './maybe-upload-media';
-import MaybeAttachMedia from './maybe-attach-media';
 
 function PostPublishPanelPrepublish( { children } ) {
 	const {
@@ -103,7 +102,6 @@ function PostPublishPanelPrepublish( { children } ) {
 				</div>
 			</div>
 			<MaybeUploadMedia />
-			<MaybeAttachMedia />
 			{ hasPublishAction && (
 				<>
 					<PanelBody